DLS-1393 GMP Bootstrap Upgrade - Frontend(catalogue update)
Updated as per comments
