+++ app/controllers/BulkReferenceController.scala
@@ -17,18 +17,18 @@
 package controllers
 
 import com.google.inject.{Inject, Singleton}
-import config.GmpFrontendAuditConnector
 import controllers.auth.AuthAction
 import forms.BulkReferenceForm
 import play.api.Logger
 import uk.gov.hmrc.auth.core.AuthConnector
+import uk.gov.hmrc.play.audit.http.connector.AuditConnector
 
 import scala.concurrent.Future
 
 @Singleton
 class BulkReferenceController @Inject()(authAction: AuthAction,
                                         val authConnector: AuthConnector,
-                                        auditConnector : GmpFrontendAuditConnector) extends GmpController {
+                                        auditConnector : AuditConnector) extends GmpController {
 
   def get = authAction.async {
       implicit request =>  Future.successful(Ok(views.html.bulk_reference(BulkReferenceForm.bulkReferenceForm)))
