+++ test/controllers/ResultsControllerSpec.scala
@@ -19,7 +19,7 @@ package controllers
 import java.util.UUID
 
 import config.ApplicationConfig.globalErrors
-import config.{ApplicationConfig, GmpFrontendAuditConnector}
+import config.{ApplicationConfig}
 import connectors.GmpConnector
 import controllers.auth.{AuthAction, FakeAuthAction, GmpAuthConnector}
 import helpers.RandomNino
@@ -40,7 +40,7 @@ import services.SessionService
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.http.logging.SessionId
-import uk.gov.hmrc.play.audit.http.connector.AuditResult
+import uk.gov.hmrc.play.audit.http.connector.{AuditConnector, AuditResult}
 import views.helpers.GmpDateFormatter._
 
 import scala.concurrent.Future
@@ -51,7 +51,7 @@ class ResultsControllerSpec extends PlaySpec with OneServerPerSuite with Mockito
   val mockSessionService = mock[SessionService]
   val mockCalculationConnector = mock[GmpConnector]
   val mockApplicationConfig = mock[ApplicationConfig]
-  val mockAuditConnector = mock[GmpFrontendAuditConnector]
+  val mockAuditConnector = mock[AuditConnector]
   val mockAuthAction = mock[AuthAction]
   val metrics = app.injector.instanceOf[ApplicationMetrics]
